Moved into folders. Refactored MapScreen.kt
