--- app/src/main/java/no/hiof/mariusca/stitur/ui/screen/EditProfileScreen.kt
@@ -1,1 +1,0 @@
-package no.hiof.mariusca.stitur.ui.screen
