+++ app/src/main/java/no/hiof/mariusca/stitur/ui/screen/map/WeatherScreen.kt
@@ -1,4 +1,4 @@
-package no.hiof.mariusca.stitur.ui.screen
+package no.hiof.mariusca.stitur.ui.screen.map
 
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
@@ -8,7 +8,7 @@ import androidx.compose.ui.text.style.TextAlign
 
 @Composable
 fun WeatherScreen() {
-    Text(text = "Weather",
+    Text(text = "Weather (Not implemented yet)",
         style = MaterialTheme.typography.headlineLarge,
         textAlign = TextAlign.Center)
 }
