+++ app/src/main/java/no/hiof/mariusca/stitur/ui/screen/map/dialogs/SaveGeoTreasureDialog.kt
@@ -28,7 +28,7 @@ import no.hiof.mariusca.stitur.model.GeoTreasure
 import no.hiof.mariusca.stitur.model.MinimalProfile
 import no.hiof.mariusca.stitur.signup.SignUpViewModel
 import no.hiof.mariusca.stitur.ui.screen.map.GeoTreasureViewModel
-import no.hiof.mariusca.stitur.ui.screen.ProfileViewModel
+import no.hiof.mariusca.stitur.ui.screen.profile.ProfileViewModel
 
 @Composable
 fun SaveGeoTreasureDialog(
