+++ app/src/main/java/no/hiof/mariusca/stitur/ui/screen/map/search/ColumnItem.kt
@@ -1,0 +1,23 @@
+package no.hiof.mariusca.stitur.ui.screen.map.search
+
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Divider
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+
+@Composable
+fun ColumnItem(item: String, onItemClick: () -> Unit) {
+    Column(
+        modifier = Modifier
+            .clickable(onClick = onItemClick)
+            .padding(10.dp)
+    ) {
+        Text(text = item, modifier = Modifier.padding(vertical = 10.dp), fontSize = 22.sp)
+        Divider()
+    }
+}
